+++ src/Routes/routesName.js
@@ -1,6 +1,6 @@
 export const ROUTES = {
   mainPage: "/",
-  cardPage: (id) => (id ? `gallery/${id}` : "gallery/id"),
+  cardPage: (id) => `gallery/${id}`,
   profilePage: "/profile",
   statisticPage: "/statistic",
 };
